requirements.txt
anyio==3.7.0
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
comm==0.1.4
coverage==7.4.4
cryptography==41.0.1
debugpy==1.6.7.post1
decorator==5.1.1
Deprecated==1.2.14
ecdsa==0.18.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.96.0
firebase-admin==6.2.0
gcloud==0.18.3
google-api-core==2.12.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==2.0.2
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.12.3
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jwcrypto==1.5.0
matplotlib-inline==0.1.6
msgpack==1.0.7
nest-asyncio==1.5.7
oauth2client==4.1.3
packaging==23.1
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
platformdirs==3.10.0
pluggy==1.3.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.25.0
psutil==5.9.5
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.8
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==5.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-jwt==4.0.0
python-multipart==0.0.6
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
stack-data==0.6.2
starlette==0.27.0
stripe==7.3.0
tomli==2.0.1
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.22.0
watchfiles==0.21.0
wcwidth==0.2.6
websockets==11.0.3
wrapt==1.15.0
